requirements.txt
transformers==4.6.0
progressbar2
numpy==1.18.0
pathlib2
torch==1.6.0
torchvision==0.7.0
tqdm
tensorboard
tensorboardX
psutil==5.6.6
scipy>=1.4.1
youtokentome==1.0.6
